refactor: rename middleware to proxy for next.js 16 compatibility
