+++ turbo/knip.json
@@ -6,8 +6,8 @@
         "app/**/page.{ts,tsx}",
         "app/**/layout.{ts,tsx}",
         "app/api/**/route.{ts,tsx}",
-        "middleware.{ts,tsx}",
-        "middleware.cors.ts",
+        "proxy.{ts,tsx}",
+        "proxy.cors.ts",
         "i18n.ts",
         "navigation.ts",
         "scripts/**/*.ts",
